preliminary refactoring
