+++ iris/src/main/java/cfa/vo/iris/smoketest/SherpaSmokeTest.java
@@ -26,6 +26,13 @@ import cfa.vo.interop.SampService;
 import cfa.vo.iris.IrisApplication;
 import cfa.vo.iris.interop.AbstractSedMessageHandler;
 import cfa.vo.iris.sed.ExtSed;
+import cfa.vo.sherpa.models.AbstractModel;
+import cfa.vo.sherpa.models.CompositeModel;
+import cfa.vo.sherpa.models.Models;
+import cfa.vo.sherpa.optimization.Method;
+import cfa.vo.sherpa.optimization.OptimizationMethod;
+import cfa.vo.sherpa.stats.Stat;
+import cfa.vo.sherpa.stats.Stats;
 import cfa.vo.utils.Default;
 import cfa.vo.sed.builder.SegmentImporter;
 import cfa.vo.sed.builder.photfilters.EnergyBin;
