+++ samp-factory/src/main/java/cfa/vo/sherpa/optimization/Method.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012, 2015 Smithsonian Astrophysical Observatory
+ * Copyright (C) 2012 Smithsonian Astrophysical Observatory
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package cfa.vo.sherpa;
+package cfa.vo.sherpa.optimization;
 
-import java.util.Arrays;
+import cfa.vo.sherpa.Config;
 
 /**
  *
  * @author olaurino
  */
-public class Polynom1D extends AbstractModel implements Model {
+public interface Method {
+    String getName();
 
-    public Polynom1D(String id) {
-        super("polynom1d", id);
-        String[] parNames = new String[]{"c0", "c1", "c2", "c3", "c4", "c5", "c6", "c7", "c8", "offset"};
-        addParams(Arrays.asList(parNames));
-    }
+    void setName(String name);
+
+    Config getConfig();
 
 }
