+++ samp-factory/src/main/java/cfa/vo/sherpa/optimization/NelderMead.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2012, 2015 Smithsonian Astrophysical Observatory
+ * Copyright (C) 2012 Smithsonian Astrophysical Observatory
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,12 @@
  * limitations under the License.
  */
 
-package cfa.vo.sherpa;
-
-import java.util.List;
+package cfa.vo.sherpa.optimization;
 
 /**
  *
  * @author olaurino
  */
-public interface CompositeModel {
-
-    String getName();
-
-    void setName(String name);
-
-    List<Model> getParts();
-
-    void addPart(Model model);
+public interface NelderMead extends Method {
+    
 }
