+++ samp-factory/src/main/java/cfa/vo/sherpa/stats/Stats.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cfa.vo.sherpa;
+package cfa.vo.sherpa.stats;
 
 /**
  *
