+++ sed-builder/src/main/java/cfa/vo/sed/science/ModelViewerFrame.java
@@ -33,10 +33,10 @@ import cfa.vo.iris.events.SedListener;
 import cfa.vo.iris.gui.GUIUtils;
 import cfa.vo.iris.sed.ExtSed;
 import cfa.vo.iris.sed.fit.IFit;
-import cfa.vo.sherpa.CompositeModel;
-import cfa.vo.sherpa.CompositeModelTreeModel;
+import cfa.vo.sherpa.models.CompositeModel;
+import cfa.vo.sherpa.models.CompositeModelTreeModel;
 import cfa.vo.sherpa.Parameter;
-import cfa.vo.sherpa.UserModel;
+import cfa.vo.sherpa.models.UserModel;
 import javax.swing.*;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.TreePath;
